bug fixes
fix #1: my preventing spam snag/stow code was bad. It now accounts for named weapons being mixed into a multisnag and is hopefully generally more hardy. I added a try: except: so that if in the future an error gets thrown for whatever new reasons, the chests dont also become inaccessible because of it.

When u get pushed off the cliffs, cosmetic item props werent fixed, so when cosmetics were reeled up they'd be automatically adorned because adorned=True. The same was true of stuff that had slimeoid=True, but i decided to make slimeoids not lose their cosmetics when the *player* jumps off the cliff rather than drop them and set that to false. Some slimeoids can fly and if they jumped anyway, they should probably die and turn to a heart (but thisd be such a killjoy 0_0) so i dont think it makes much sense.