Removed accidental duplicate items.
Table, Chair, Couch, Desk, and Lamp have been added to Based Hardware as easy purchases. Despite this only the Table, Chair, Couch, and Bed make up the Default set. This is to make the flavortext of the whole set more attainable with limited apartment space. (Desk and Lamp were excluded due to not being essential for a living space.)